+++ src/cmd/go/testdata/script/test_fuzz_cache.txt
@@ -1,6 +1,9 @@
 # TODO(jayconrod): support shared memory on more platforms.
 [!darwin] [!linux] [!windows] skip
 
+# TODO(#48504): fix and re-enable.
+[linux] [386] skip
+
 [short] skip
 env GOCACHE=$WORK/cache
