+++ src/internal/fuzz/coverage.go
@@ -6,27 +6,8 @@ package fuzz
 
 import (
 	"fmt"
-	"internal/unsafeheader"
 	"math/bits"
-	"unsafe"
 )
-
-// coverage returns a []byte containing unique 8-bit counters for each edge of
-// the instrumented source code. This coverage data will only be generated if
-// `-d=libfuzzer` is set at build time. This can be used to understand the code
-// coverage of a test execution.
-func coverage() []byte {
-	addr := unsafe.Pointer(&_counters)
-	size := uintptr(unsafe.Pointer(&_ecounters)) - uintptr(addr)
-
-	var res []byte
-	*(*unsafeheader.Slice)(unsafe.Pointer(&res)) = unsafeheader.Slice{
-		Data: addr,
-		Len:  int(size),
-		Cap:  int(size),
-	}
-	return res
-}
 
 // ResetCovereage sets all of the counters for each edge of the instrumented
 // source code to 0.
